Split dns settings from vpn settings in UI
