+++ ios/MullvadVPN/View controllers/Preferences/PreferencesDataSource.swift
@@ -13,104 +13,53 @@ final class PreferencesDataSource: UITableViewDiffableDataSource<
     PreferencesDataSource.Section,
     PreferencesDataSource.Item
 >, UITableViewDelegate {
-    typealias InfoButtonHandler = (PreferencesDataSource.Item) -> Void
+    typealias InfoButtonHandler = (Item) -> Void
 
     enum CellReuseIdentifiers: String, CaseIterable {
-        case setting
-        case settingSwitch
-        case dnsServer
-        case dnsServerInfo
-        case addDNSServer
+        case dnsSettings
         case wireGuardPort
         case wireGuardCustomPort
-        #if DEBUG
         case wireGuardObfuscation
         case wireGuardObfuscationPort
-        #endif
         var reusableViewClass: AnyClass {
             switch self {
-            case .setting:
+            case .dnsSettings:
                 return SettingsCell.self
-            case .settingSwitch:
-                return SettingsSwitchCell.self
-            case .dnsServer:
-                return SettingsDNSTextCell.self
-            case .dnsServerInfo:
-                return SettingsDNSInfoCell.self
-            case .addDNSServer:
-                return SettingsAddDNSEntryCell.self
             case .wireGuardPort:
                 return SelectableSettingsCell.self
             case .wireGuardCustomPort:
                 return SettingsInputCell.self
-            #if DEBUG
             case .wireGuardObfuscation:
                 return SelectableSettingsCell.self
             case .wireGuardObfuscationPort:
                 return SelectableSettingsCell.self
-            #endif
             }
         }
     }
 
     private enum HeaderFooterReuseIdentifiers: String, CaseIterable {
-        case contentBlockerHeader
         case wireGuardPortHeader
-        case spacer
 
         var reusableViewClass: AnyClass {
-            switch self {
-            case .contentBlockerHeader, .wireGuardPortHeader:
-                return SettingsHeaderView.self
-            case .spacer:
-                return EmptyTableViewHeaderFooterView.self
-            }
+            return SettingsHeaderView.self
         }
     }
 
-    enum InfoButtonItem {
-        case contentBlockers
-        case blockMalware
-        case wireGuardPorts
-        #if DEBUG
-        case wireGuardObfuscation
-        case wireGuardObfuscationPort
-        #endif
-    }
-
     enum Section: String, Hashable, CaseIterable {
-        case contentBlockers
-        case customDNS
+        case dnsSettings
         case wireGuardPorts
-        #if DEBUG
         case wireGuardObfuscation
         case wireGuardObfuscationPort
-        #endif
     }
 
     enum Item: Hashable {
-        case blockAdvertising
-        case blockTracking
-        case blockMalware
-        case blockAdultContent
-        case blockGambling
-        case blockSocialMedia
+        case dnsSettings
         case wireGuardPort(_ port: UInt16?)
         case wireGuardCustomPort
-        case useCustomDNS
-        case addDNSServer
-        case dnsServer(_ uniqueID: UUID)
-        case dnsServerInfo
-        #if DEBUG
         case wireGuardObfuscationAutomatic
         case wireGuardObfuscationOn
         case wireGuardObfuscationOff
         case wireGuardObfuscationPort(_ port: UInt16)
-        #endif
-
-        static var contentBlockers: [Item] {
-            [.blockAdvertising, .blockTracking, .blockMalware, .blockAdultContent, .blockGambling, .blockSocialMedia]
-        }
 
         static var wireGuardPorts: [Item] {
             let defaultPorts = PreferencesViewModel.defaultWireGuardPorts.map {
@@ -119,7 +68,6 @@ final class PreferencesDataSource: UITableViewDiffableDataSource<
             return [.wireGuardPort(nil)] + defaultPorts + [.wireGuardCustomPort]
         }
 
-        #if DEBUG
         static var wireGuardObfuscation: [Item] {
             [.wireGuardObfuscationAutomatic, .wireGuardObfuscationOn, wireGuardObfuscationOff]
         }
@@ -127,21 +75,11 @@ final class PreferencesDataSource: UITableViewDiffableDataSource<
         static var wireGuardObfuscationPort: [Item] {
             [.wireGuardObfuscationPort(0), wireGuardObfuscationPort(80), wireGuardObfuscationPort(5001)]
         }
-        #endif
+
         var accessibilityIdentifier: String {
             switch self {
-            case .blockAdvertising:
-                return "blockAdvertising"
-            case .blockTracking:
-                return "blockTracking"
-            case .blockMalware:
-                return "blockMalware"
-            case .blockGambling:
-                return "blockGambling"
-            case .blockAdultContent:
-                return "blockAdultContent"
-            case .blockSocialMedia:
-                return "blockSocialMedias"
+            case .dnsSettings:
+                return "DNS settings"
             case let .wireGuardPort(port):
                 if let port {
                     return "wireGuardPort(\(port))"
@@ -150,15 +88,6 @@ final class PreferencesDataSource: UITableViewDiffableDataSource<
                 }
             case .wireGuardCustomPort:
                 return "wireGuardCustomPort"
-            case .useCustomDNS:
-                return "useCustomDNS"
-            case .addDNSServer:
-                return "addDNSServer"
-            case let .dnsServer(uuid):
-                return "dnsServer(\(uuid.uuidString))"
-            case .dnsServerInfo:
-                return "dnsServerInfo"
-            #if DEBUG
             case .wireGuardObfuscationAutomatic:
                 return "Automatic"
             case .wireGuardObfuscationOn:
@@ -170,59 +99,51 @@ final class PreferencesDataSource: UITableViewDiffableDataSource<
                     return "Automatic"
                 }
                 return "\(port)"
-            #endif
             }
         }
 
-        static func isDNSServerItem(_ item: Item) -> Bool {
-            if case .dnsServer = item {
-                return true
-            } else {
-                return false
-            }
-        }
-
-        var reuseIdentifier: PreferencesDataSource.CellReuseIdentifiers {
+        var reuseIdentifier: CellReuseIdentifiers {
             switch self {
-            case .addDNSServer:
-                return .addDNSServer
-            case .dnsServer:
-                return .dnsServer
-            case .dnsServerInfo:
-                return .dnsServerInfo
+            case .dnsSettings:
+                return .dnsSettings
             case .wireGuardPort:
                 return .wireGuardPort
             case .wireGuardCustomPort:
                 return .wireGuardCustomPort
-            #if DEBUG
             case .wireGuardObfuscationAutomatic, .wireGuardObfuscationOn, .wireGuardObfuscationOff:
                 return .wireGuardObfuscation
             case .wireGuardObfuscationPort:
                 return .wireGuardObfuscationPort
-            #endif
-            default:
-                return .settingSwitch
             }
         }
     }
 
-    private var isEditing = false
-
     private(set) var viewModel = PreferencesViewModel() { didSet {
         preferencesCellFactory.viewModel = viewModel
     }}
-    private(set) var viewModelBeforeEditing = PreferencesViewModel()
     private let preferencesCellFactory: PreferencesCellFactory
     private weak var tableView: UITableView?
 
     weak var delegate: PreferencesDataSourceDelegate?
 
-    var indexPathForSelectedPort: IndexPath? {
-        let selectedItem: Item = viewModel.customWireGuardPort == nil
+    var selectedIndexPaths: [IndexPath] {
+        let wireGuardPortItem: Item = viewModel.customWireGuardPort == nil
             ? .wireGuardPort(viewModel.wireGuardPort)
             : .wireGuardCustomPort
 
-        return indexPath(for: selectedItem)
+        let obfuscationStateItem: Item = switch viewModel.obfuscationState {
+        case .automatic: .wireGuardObfuscationAutomatic
+        case .off: .wireGuardObfuscationOff
+        case .on: .wireGuardObfuscationOn
+        }
+
+        let obfuscationPortItem: Item = .wireGuardObfuscationPort(viewModel.obfuscationPort.portValue)
+
+        return [
+            indexPath(for: wireGuardPortItem),
+            indexPath(for: obfuscationStateItem),
+            indexPath(for: obfuscationPortItem),
+        ].compactMap { $0 }
     }
 
     init(tableView: UITableView) {
@@ -248,31 +169,6 @@ final class PreferencesDataSource: UITableViewDiffableDataSource<
         viewModel.availableWireGuardPortRanges = ranges
     }
 
-    func setEditing(_ editing: Bool, animated: Bool) {
-        guard isEditing != editing else { return }
-
-        isEditing = editing
-        preferencesCellFactory.isEditing = isEditing
-
-        if editing {
-            viewModelBeforeEditing = viewModel
-        } else {
-            viewModel.sanitizeCustomDNSEntries()
-        }
-
-        updateSnapshot(animated: true)
-
-        viewModel.customDNSDomains.forEach { entry in
-            self.reload(item: .dnsServer(entry.identifier))
-        }
-
-        if !editing, viewModelBeforeEditing != viewModel {
-            delegate?.preferencesDataSource(self, didChangeViewModel: viewModel)
-        }
-
-        selectRow(at: indexPathForSelectedPort)
-    }
-
     func revertWireGuardPortCellToLastSelection() {
         guard let customPortCell = getCustomPortCell(), customPortCell.textField.isEditing else {
             return
@@ -291,7 +187,6 @@ final class PreferencesDataSource: UITableViewDiffableDataSource<
             Item.wireGuardPorts.forEach { item in
                 if case let .wireGuardPort(port) = item, port == viewModel.wireGuardPort {
                     selectRow(at: item)
-
                     return
                 }
             }
@@ -309,93 +204,11 @@ final class PreferencesDataSource: UITableViewDiffableDataSource<
         updateSnapshot()
     }
 
-    // MARK: - UITableViewDataSource
-
-    func tableView(
-        _ tableView: UITableView,
-        willDisplay cell: UITableViewCell,
-        forRowAt indexPath: IndexPath
-    ) {
-        if indexPath == indexPathForSelectedPort {
-            cell.setSelected(true, animated: false)
-        }
-    }
-
-    override func tableView(_ tableView: UITableView, canEditRowAt indexPath: IndexPath) -> Bool {
-        // Disable swipe to delete when not editing the table view
-        guard isEditing else { return false }
-
-        let item = itemIdentifier(for: indexPath)
-
-        switch item {
-        case .dnsServer, .addDNSServer:
-            return true
-        default:
-            return false
-        }
-    }
-
-    override func tableView(
-        _ tableView: UITableView,
-        commit editingStyle: UITableViewCell.EditingStyle,
-        forRowAt indexPath: IndexPath
-    ) {
-        let item = itemIdentifier(for: indexPath)
-
-        if case .addDNSServer = item, editingStyle == .insert {
-            addDNSServerEntry()
-        }
-
-        if case let .dnsServer(entryIdentifier) = item, editingStyle == .delete {
-            deleteDNSServerEntry(entryIdentifier: entryIdentifier)
-        }
-    }
-
-    override func tableView(_ tableView: UITableView, canMoveRowAt indexPath: IndexPath) -> Bool {
-        let item = itemIdentifier(for: indexPath)
-
-        switch item {
-        case .dnsServer:
-            return true
-        default:
-            return false
-        }
-    }
-
-    override func tableView(
-        _ tableView: UITableView,
-        moveRowAt sourceIndexPath: IndexPath,
-        to destinationIndexPath: IndexPath
-    ) {
-        let sourceItem = itemIdentifier(for: sourceIndexPath)!
-        let destinationItem = itemIdentifier(for: destinationIndexPath)!
-
-        guard case let .dnsServer(sourceIdentifier) = sourceItem,
-              case let .dnsServer(targetIdentifier) = destinationItem,
-              let sourceIndex = viewModel.indexOfDNSEntry(entryIdentifier: sourceIdentifier),
-              let destinationIndex = viewModel.indexOfDNSEntry(entryIdentifier: targetIdentifier)
-        else { return }
-
-        let removedEntry = viewModel.customDNSDomains.remove(at: sourceIndex)
-        viewModel.customDNSDomains.insert(removedEntry, at: destinationIndex)
-
-        updateSnapshot()
-    }
-
     // MARK: - UITableViewDelegate
 
-    func tableView(_ tableView: UITableView, shouldHighlightRowAt indexPath: IndexPath) -> Bool {
-        switch itemIdentifier(for: indexPath) {
-        #if DEBUG
-        case .wireGuardPort, .wireGuardCustomPort, .wireGuardObfuscationAutomatic, .wireGuardObfuscationOn,
-             .wireGuardObfuscationOff, .wireGuardObfuscationPort:
-            return true
-        #else
-        case .wireGuardPort, .wireGuardCustomPort:
-            return true
-        #endif
-        default:
-            return false
+    func tableView(_ tableView: UITableView, willDisplay cell: UITableViewCell, forRowAt indexPath: IndexPath) {
+        if selectedIndexPaths.contains(indexPath) {
+            cell.setSelected(true, animated: false)
         }
     }
 
@@ -405,6 +218,10 @@ final class PreferencesDataSource: UITableViewDiffableDataSource<
         deselectAllRowsInSectionExceptRowAt(indexPath)
 
         switch item {
+        case .dnsSettings:
+            tableView.deselectRow(at: indexPath, animated: false)
+            delegate?.showDNSSettings()
+
         case let .wireGuardPort(port):
             viewModel.setWireGuardPort(port)
 
@@ -413,28 +230,39 @@ final class PreferencesDataSource: UITableViewDiffableDataSource<
                 cell.textField.resignFirstResponder()
             }
 
-            delegate?.preferencesDataSource(self, didSelectPort: port)
+            delegate?.didSelectWireGuardPort(port)
 
         case .wireGuardCustomPort:
             getCustomPortCell()?.textField.becomeFirstResponder()
 
-        #if DEBUG
         case .wireGuardObfuscationAutomatic:
-            print("UDP over TCP Set to automatic")
+            selectObfuscationState(.automatic)
+            delegate?.didChangeViewModel(viewModel)
         case .wireGuardObfuscationOn:
-            print("turning on UDP over TCP")
+            selectObfuscationState(.on)
+            delegate?.didChangeViewModel(viewModel)
         case .wireGuardObfuscationOff:
-            print("Turning off UDP over TCP")
+            selectObfuscationState(.off)
+            delegate?.didChangeViewModel(viewModel)
         case let .wireGuardObfuscationPort(port):
-            print("Setting port to \(port)")
-        #endif
+            selectObfuscationPort(port)
+            delegate?.didChangeViewModel(viewModel)
+
         default:
             break
         }
     }
 
+    // Disallow selection for tapping on a selectable setting.
     func tableView(_ tableView: UITableView, willDeselectRowAt indexPath: IndexPath) -> IndexPath? {
-        nil
+        let item = itemIdentifier(for: indexPath)
+
+        switch item {
+        case .dnsSettings:
+            return indexPath
+        default:
+            return nil
+        }
     }
 
     func tableView(_ tableView: UITableView, viewForHeaderInSection section: Int) -> UIView? {
@@ -442,49 +270,36 @@ final class PreferencesDataSource: UITableViewDiffableDataSource<
 
         guard let view = tableView
             .dequeueReusableHeaderFooterView(
-                withIdentifier: HeaderFooterReuseIdentifiers.contentBlockerHeader
+                withIdentifier: HeaderFooterReuseIdentifiers.wireGuardPortHeader
                     .rawValue
             ) as? SettingsHeaderView else { return nil }
 
         switch sectionIdentifier {
-        case .contentBlockers:
-            configureContentBlockersHeader(view)
-            return view
         case .wireGuardPorts:
             configureWireguardPortsHeader(view)
             return view
 
-        #if DEBUG
         case .wireGuardObfuscation:
             configureObfuscationHeader(view)
             return view
         case .wireGuardObfuscationPort:
             configureObfuscationPortHeader(view)
             return view
-        #endif
+
         default:
             return nil
         }
     }
 
     func tableView(_ tableView: UITableView, viewForFooterInSection section: Int) -> UIView? {
-        let sectionIdentifier = snapshot().sectionIdentifiers[section]
-
-        switch sectionIdentifier {
-        case .wireGuardPorts:
-            return tableView.dequeueReusableHeaderFooterView(
-                withIdentifier: HeaderFooterReuseIdentifiers.spacer.rawValue
-            )
-        default:
-            return nil
-        }
+        nil
     }
 
     func tableView(_ tableView: UITableView, heightForHeaderInSection section: Int) -> CGFloat {
         let sectionIdentifier = snapshot().sectionIdentifiers[section]
 
         switch sectionIdentifier {
-        case .customDNS:
+        case .dnsSettings:
             return 0
 
         default:
@@ -496,64 +311,10 @@ final class PreferencesDataSource: UITableViewDiffableDataSource<
         let sectionIdentifier = snapshot().sectionIdentifiers[section]
 
         switch sectionIdentifier {
-        #if DEBUG
-        case .wireGuardObfuscationPort:
-            return UIMetrics.TableView.sectionSpacing
-        #else
-        case .wireGuardPorts:
-            return UIMetrics.TableView.sectionSpacing
-        #endif
-
+        case .dnsSettings:
+            return 10
         default:
-            return 0
-        }
-    }
-
-    func tableView(
-        _ tableView: UITableView,
-        editingStyleForRowAt indexPath: IndexPath
-    ) -> UITableViewCell.EditingStyle {
-        let item = itemIdentifier(for: indexPath)
-
-        switch item {
-        case .dnsServer:
-            return .delete
-        case .addDNSServer:
-            return .insert
-        default:
-            return .none
-        }
-    }
-
-    func tableView(
-        _ tableView: UITableView,
-        targetIndexPathForMoveFromRowAt sourceIndexPath: IndexPath,
-        toProposedIndexPath proposedDestinationIndexPath: IndexPath
-    ) -> IndexPath {
-        let sectionIdentifier = snapshot().sectionIdentifiers[sourceIndexPath.section]
-        guard case .customDNS = sectionIdentifier else { return sourceIndexPath }
-
-        let items = snapshot().itemIdentifiers(inSection: sectionIdentifier)
-
-        let indexPathForFirstRow = items.first(where: Item.isDNSServerItem).flatMap { item in
-            indexPath(for: item)
-        }
-
-        let indexPathForLastRow = items.last(where: Item.isDNSServerItem).flatMap { item in
-            indexPath(for: item)
-        }
-
-        guard let indexPathForFirstRow,
-              let indexPathForLastRow else { return sourceIndexPath }
-
-        if proposedDestinationIndexPath.section == sourceIndexPath.section {
-            return min(max(proposedDestinationIndexPath, indexPathForFirstRow), indexPathForLastRow)
-        } else {
-            if proposedDestinationIndexPath.section > sourceIndexPath.section {
-                return indexPathForLastRow
-            } else {
-                return indexPathForFirstRow
-            }
+            return 0.5
         }
     }
 
@@ -567,12 +328,10 @@ final class PreferencesDataSource: UITableViewDiffableDataSource<
             )
         }
 
-        HeaderFooterReuseIdentifiers.allCases.forEach { enumCase in
-            tableView?.register(
-                enumCase.reusableViewClass,
-                forHeaderFooterViewReuseIdentifier: enumCase.rawValue
-            )
-        }
+        tableView?.register(
+            SettingsHeaderView.self,
+            forHeaderFooterViewReuseIdentifier: HeaderFooterReuseIdentifiers.wireGuardPortHeader.rawValue
+        )
     }
 
     private func deselectAllRowsInSectionExceptRowAt(_ indexPath: IndexPath) {
@@ -585,53 +344,12 @@ final class PreferencesDataSource: UITableViewDiffableDataSource<
     }
 
     private func updateSnapshot(animated: Bool = false, completion: (() -> Void)? = nil) {
-        var newSnapshot = NSDiffableDataSourceSnapshot<Section, Item>()
-        let oldSnapshot = snapshot()
+        var snapshot = NSDiffableDataSourceSnapshot<Section, Item>()
 
-        newSnapshot.appendSections(Section.allCases)
+        snapshot.appendSections(Section.allCases)
+        snapshot.appendItems([.dnsSettings], toSection: .dnsSettings)
 
-        // Append sections
-
-        if oldSnapshot.sectionIdentifiers.contains(.contentBlockers) {
-            newSnapshot.appendItems(
-                oldSnapshot.itemIdentifiers(inSection: .contentBlockers),
-                toSection: .contentBlockers
-            )
-        }
-
-        if oldSnapshot.sectionIdentifiers.contains(.wireGuardPorts) {
-            newSnapshot.appendItems(
-                oldSnapshot.itemIdentifiers(inSection: .wireGuardPorts),
-                toSection: .wireGuardPorts
-            )
-        }
-
-        // Append DNS settings
-
-        newSnapshot.appendItems([.useCustomDNS], toSection: .customDNS)
-
-        let dnsServerItems = viewModel.customDNSDomains.map { entry in
-            Item.dnsServer(entry.identifier)
-        }
-        newSnapshot.appendItems(dnsServerItems, toSection: .customDNS)
-
-        if isEditing, viewModel.customDNSDomains.count < DNSSettings.maxAllowedCustomDNSDomains {
-            newSnapshot.appendItems([.addDNSServer], toSection: .customDNS)
-        }
-
-        // Append/update DNS server info.
-
-        if viewModel.customDNSPrecondition == .satisfied {
-            newSnapshot.deleteItems([.dnsServerInfo])
-        } else {
-            if newSnapshot.itemIdentifiers(inSection: .customDNS).contains(.dnsServerInfo) {
-                newSnapshot.reloadItems([.dnsServerInfo])
-            } else {
-                newSnapshot.appendItems([.dnsServerInfo], toSection: .customDNS)
-            }
-        }
-
-        applySnapshot(newSnapshot, animated: animated, completion: completion)
+        applySnapshot(snapshot, animated: animated, completion: completion)
     }
 
     private func applySnapshot(
@@ -640,7 +358,7 @@ final class PreferencesDataSource: UITableViewDiffableDataSource<
         completion: (() -> Void)? = nil
     ) {
         apply(snapshot, animatingDifferences: animated) { [weak self] in
-            self?.selectRow(at: self?.indexPathForSelectedPort)
+            self?.selectedIndexPaths.forEach { self?.selectRow(at: $0) }
             completion?()
         }
     }
@@ -649,205 +367,6 @@ final class PreferencesDataSource: UITableViewDiffableDataSource<
         if let indexPath = indexPath(for: item),
            let cell = tableView?.cellForRow(at: indexPath) {
             preferencesCellFactory.configureCell(cell, item: item, indexPath: indexPath)
-        }
-    }
-
-    private func setBlockAdvertising(_ isEnabled: Bool) {
-        let oldViewModel = viewModel
-
-        viewModel.setBlockAdvertising(isEnabled)
-
-        if oldViewModel.customDNSPrecondition != viewModel.customDNSPrecondition {
-            reloadDnsServerInfo()
-        }
-
-        if !isEditing {
-            delegate?.preferencesDataSource(self, didChangeViewModel: viewModel)
-        }
-    }
-
-    private func setBlockTracking(_ isEnabled: Bool) {
-        let oldViewModel = viewModel
-
-        viewModel.setBlockTracking(isEnabled)
-
-        if oldViewModel.customDNSPrecondition != viewModel.customDNSPrecondition {
-            reloadDnsServerInfo()
-        }
-
-        if !isEditing {
-            delegate?.preferencesDataSource(self, didChangeViewModel: viewModel)
-        }
-    }
-
-    private func setBlockMalware(_ isEnabled: Bool) {
-        let oldViewModel = viewModel
-
-        viewModel.setBlockMalware(isEnabled)
-
-        if oldViewModel.customDNSPrecondition != viewModel.customDNSPrecondition {
-            reloadDnsServerInfo()
-        }
-
-        if !isEditing {
-            delegate?.preferencesDataSource(self, didChangeViewModel: viewModel)
-        }
-    }
-
-    private func setBlockAdultContent(_ isEnabled: Bool) {
-        let oldViewModel = viewModel
-
-        viewModel.setBlockAdultContent(isEnabled)
-
-        if oldViewModel.customDNSPrecondition != viewModel.customDNSPrecondition {
-            reloadDnsServerInfo()
-        }
-
-        if !isEditing {
-            delegate?.preferencesDataSource(self, didChangeViewModel: viewModel)
-        }
-    }
-
-    private func setBlockGambling(_ isEnabled: Bool) {
-        let oldViewModel = viewModel
-
-        viewModel.setBlockGambling(isEnabled)
-
-        if oldViewModel.customDNSPrecondition != viewModel.customDNSPrecondition {
-            reloadDnsServerInfo()
-        }
-
-        if !isEditing {
-            delegate?.preferencesDataSource(self, didChangeViewModel: viewModel)
-        }
-    }
-
-    private func setBlockSocialMedia(_ isEnabled: Bool) {
-        let oldViewModel = viewModel
-
-        viewModel.setBlockSocialMedia(isEnabled)
-
-        if oldViewModel.customDNSPrecondition != viewModel.customDNSPrecondition {
-            reloadDnsServerInfo()
-        }
-
-        if !isEditing {
-            delegate?.preferencesDataSource(self, didChangeViewModel: viewModel)
-        }
-    }
-
-    private func setEnableCustomDNS(_ isEnabled: Bool) {
-        let oldViewModel = viewModel
-
-        viewModel.setEnableCustomDNS(isEnabled)
-
-        if oldViewModel.customDNSPrecondition != viewModel.customDNSPrecondition {
-            reloadDnsServerInfo()
-        }
-
-        if !isEditing {
-            delegate?.preferencesDataSource(self, didChangeViewModel: viewModel)
-        }
-    }
-
-    private func handleDNSEntryChange(with identifier: UUID, inputString: String) -> Bool {
-        let oldViewModel = viewModel
-
-        viewModel.updateDNSEntry(entryIdentifier: identifier, newAddress: inputString)
-
-        if oldViewModel.customDNSPrecondition != viewModel.customDNSPrecondition {
-            reloadDnsServerInfo()
-        }
-
-        return viewModel.isDNSDomainUserInputValid(inputString)
-    }
-
-    private func addDNSServerEntry() {
-        let newDNSEntry = DNSServerEntry(address: "")
-        viewModel.customDNSDomains.append(newDNSEntry)
-
-        updateSnapshot(animated: true) { [weak self] in
-            // Focus on the new entry text field.
-            let lastDNSEntry = self?.snapshot().itemIdentifiers(inSection: .customDNS)
-                .last { item in
-                    if case let .dnsServer(entryIdentifier) = item {
-                        return entryIdentifier == newDNSEntry.identifier
-                    } else {
-                        return false
-                    }
-                }
-
-            if let lastDNSEntry,
-               let indexPath = self?.indexPath(for: lastDNSEntry) {
-                let cell = self?.tableView?.cellForRow(at: indexPath) as? SettingsDNSTextCell
-
-                self?.tableView?.scrollToRow(at: indexPath, at: .bottom, animated: true)
-                cell?.textField.becomeFirstResponder()
-            }
-        }
-    }
-
-    private func deleteDNSServerEntry(entryIdentifier: UUID) {
-        let entryIndex = viewModel.customDNSDomains.firstIndex { entry in
-            entry.identifier == entryIdentifier
-        }
-
-        guard let entryIndex else { return }
-
-        viewModel.customDNSDomains.remove(at: entryIndex)
-
-        reload(item: .useCustomDNS)
-        updateSnapshot(animated: true)
-    }
-
-    private func reloadDnsServerInfo() {
-        var snapshot = snapshot()
-
-        reload(item: .useCustomDNS)
-
-        if viewModel.customDNSPrecondition == .satisfied {
-            snapshot.deleteItems([.dnsServerInfo])
-        } else {
-            if snapshot.itemIdentifiers(inSection: .customDNS).contains(.dnsServerInfo) {
-                snapshot.reloadItems([.dnsServerInfo])
-            } else {
-                snapshot.appendItems([.dnsServerInfo], toSection: .customDNS)
-            }
-        }
-
-        apply(snapshot, animatingDifferences: true)
-    }
-
-    private func configureContentBlockersHeader(_ header: SettingsHeaderView) {
-        let title = NSLocalizedString(
-            "CONTENT_BLOCKERS_HEADER_LABEL",
-            tableName: "Preferences",
-            value: "DNS content blockers",
-            comment: ""
-        )
-
-        header.titleLabel.text = title
-        header.accessibilityCustomActionName = title
-
-        header.infoButtonHandler = { [weak self] in
-            if let self {
-                self.delegate?.preferencesDataSource(self, showInfo: .contentBlockers)
-            }
-        }
-
-        header.didCollapseHandler = { [weak self] headerView in
-            guard let self else { return }
-
-            var snapshot = self.snapshot()
-
-            if headerView.isExpanded {
-                snapshot.deleteItems(Item.contentBlockers)
-            } else {
-                snapshot.appendItems(Item.contentBlockers, toSection: .contentBlockers)
-            }
-
-            headerView.isExpanded.toggle()
-            self.apply(snapshot, animatingDifferences: true)
         }
     }
 
@@ -863,7 +382,7 @@ final class PreferencesDataSource: UITableViewDiffableDataSource<
         header.accessibilityCustomActionName = title
         header.infoButtonHandler = { [weak self] in
             if let self {
-                self.delegate?.preferencesDataSource(self, showInfo: .wireGuardPorts)
+                self.delegate?.showInfo(for: .wireGuardPorts)
             }
         }
 
@@ -888,13 +407,10 @@ final class PreferencesDataSource: UITableViewDiffableDataSource<
             // user just enough time to notice it.
             DispatchQueue.main.asyncAfter(deadline: .now() + updateTimeDelay) { [weak self] in
                 headerView.isExpanded.toggle()
-
                 self?.applySnapshot(snapshot, animated: true)
             }
         }
     }
-
-    #if DEBUG
 
     private func configureObfuscationHeader(_ header: SettingsHeaderView) {
         let title = NSLocalizedString(
@@ -920,7 +436,7 @@ final class PreferencesDataSource: UITableViewDiffableDataSource<
         }
 
         header.infoButtonHandler = { [weak self] in
-            self.map { $0.delegate?.preferencesDataSource($0, showInfo: .wireGuardObfuscation) }
+            self.map { $0.delegate?.showInfo(for: .wireGuardObfuscation) }
         }
     }
 
@@ -948,12 +464,12 @@ final class PreferencesDataSource: UITableViewDiffableDataSource<
         }
 
         header.infoButtonHandler = { [weak self] in
-            self.map { $0.delegate?.preferencesDataSource($0, showInfo: .wireGuardObfuscationPort) }
+            self.map { $0.delegate?.showInfo(for: .wireGuardObfuscationPort) }
         }
     }
-    #endif
+
     private func selectRow(at indexPath: IndexPath?, animated: Bool = false) {
-        tableView?.selectRow(at: indexPath, animated: false, scrollPosition: .none)
+        tableView?.selectRow(at: indexPath, animated: animated, scrollPosition: .none)
     }
 
     private func selectRow(at item: Item?, animated: Bool = false) {
@@ -972,60 +488,31 @@ final class PreferencesDataSource: UITableViewDiffableDataSource<
 }
 
 extension PreferencesDataSource: PreferencesCellEventHandler {
-    func didChangeState(for preference: Item, isOn: Bool) {
-        switch preference {
-        case .blockAdvertising:
-            setBlockAdvertising(isOn)
-
-        case .blockTracking:
-            setBlockTracking(isOn)
-
-        case .blockMalware:
-            setBlockMalware(isOn)
-
-        case .blockAdultContent:
-            setBlockAdultContent(isOn)
-
-        case .blockGambling:
-            setBlockGambling(isOn)
-
-        case .blockSocialMedia:
-            setBlockSocialMedia(isOn)
-
-        case .useCustomDNS:
-            setEnableCustomDNS(isOn)
-
-        default:
-            break
-        }
-    }
-
-    func addDNSEntry() {
-        addDNSServerEntry()
-    }
-
-    func didChangeDNSEntry(
-        with identifier: UUID,
-        inputString: String
-    ) -> Bool {
-        handleDNSEntryChange(with: identifier, inputString: inputString)
-    }
-
-    func showInfo(for button: InfoButtonItem) {
-        delegate?.preferencesDataSource(self, showInfo: button)
+    func showInfo(for button: PreferencesInfoButtonItem) {
+        delegate?.showInfo(for: button)
     }
 
     func addCustomPort(_ port: UInt16) {
         viewModel.setWireGuardPort(port)
-        delegate?.preferencesDataSource(self, didSelectPort: port)
+        delegate?.didSelectWireGuardPort(port)
     }
 
     func selectCustomPortEntry(_ port: UInt16) -> Bool {
-        if indexPathForSelectedPort != indexPath(for: .wireGuardCustomPort) {
-            selectRow(at: .wireGuardCustomPort)
+        if let indexPath = indexPath(for: .wireGuardCustomPort), !selectedIndexPaths.contains(indexPath) {
+            deselectAllRowsInSectionExceptRowAt(indexPath)
+            selectRow(at: indexPath)
         }
 
         return viewModel.isPortWithinValidWireGuardRanges(port)
+    }
+
+    func selectObfuscationState(_ state: WireGuardObfuscationState) {
+        viewModel.setWireGuardObfuscationState(state)
+    }
+
+    func selectObfuscationPort(_ port: UInt16) {
+        let selectedPort = WireGuardObfuscationPort(rawValue: port)!
+        viewModel.setWireGuardObfuscationPort(selectedPort)
     }
 }
