+++ ios/MullvadVPN/View controllers/Preferences/PreferencesInfoButtonItem.swift
@@ -1,0 +1,15 @@
+//
+//  PreferencesInfoButtonItem.swift
+//  MullvadVPN
+//
+//  Created by Jon Petersson on 2023-11-10.
+//  Copyright © 2023 Mullvad VPN AB. All rights reserved.
+//
+
+enum PreferencesInfoButtonItem {
+    case contentBlockers
+    case blockMalware
+    case wireGuardPorts
+    case wireGuardObfuscation
+    case wireGuardObfuscationPort
+}
